+++ private/buf/bufcurl/headers.go
