+++ private/buf/bufcurl/resolver.go
